chore(stella-tui): split deck_ui/tests.rs before it reaches the gate ceiling
The file was approaching the 1500-line ratchet with no baseline entry, so a
crossing would have failed the gate outright rather than ratcheting — the split
has to precede the growth, not follow it.

A pure move into the `deck_ui/tests/` directory that already held seven themed
modules: 54 tests leave `tests.rs` for seven new siblings (selection, sessions,
splash, composer, tabs, routing_card, esc), and `tests.rs` drops from 1370 lines
to 87 as a module declaration.

Verified as a move rather than a rewrite: `tests.rs` carried 54 `#[test]`
attributes before and carries 0 now; the seven new files carry exactly 54
between them. Nothing was changed, added or deleted, which is what
`scripts/check-deleted-tests.sh` compares the base branch against the merge
result for. The crate suite goes from 966 to 975 passing across this branch's
waves with no loss here.

No baseline entry moves: the file was never grandfathered, so this buys
headroom under the standing 1500-line limit rather than retiring an exemption.

Closes #3742
