+++ crates/stella-tui/src/deck_ui/tests.rs
@@ -10,12 +10,19 @@ use super::*;
 use crate::envelope::AgentMeta;
 use stella_protocol::AgentEvent;
 
+mod composer;
+mod esc;
 mod gates;
 mod graph;
 mod help;
 mod issues;
 mod queue;
+mod routing_card;
+mod selection;
+mod sessions;
 mod skills;
+mod splash;
+mod tabs;
 mod transcript_nav;
 
 /// A model whose lead already has `prompts` queued, for the queue-editor and
@@ -36,14 +43,6 @@ fn ch(c: char) -> KeyEvent {
 }
 fn ctrl(c: char) -> KeyEvent {
     KeyEvent::new(KeyCode::Char(c), KeyModifiers::CONTROL)
-}
-/// The newline chord — `⌘⏎` as the kitty keyboard protocol reports it
-/// (a modified Enter inserts a line break; a bare Enter submits).
-fn cmd_enter() -> KeyEvent {
-    KeyEvent::new(KeyCode::Enter, KeyModifiers::SUPER)
-}
-fn alt(c: char) -> KeyEvent {
-    KeyEvent::new(KeyCode::Char(c), KeyModifiers::ALT)
 }
 
 fn model_with(ids: &[&str]) -> WorkspaceModel {
@@ -81,1290 +80,8 @@ fn with_tool_exchange(m: &mut WorkspaceModel, agent: &str) {
     });
 }
 
-#[test]
-fn up_selects_the_last_message_and_ctrl_o_toggles_it() {
-    let mut model = model_with(&["lead"]);
-    with_tool_exchange(&mut model, "lead");
-    let mut ui = ready_ui();
-
-    handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-    assert_eq!(
-        ui.session_selected,
-        Some(1),
-        "up highlights the newest message"
-    );
-
-    handle_deck_key(ctrl('o'), &model, &mut ui);
-    assert!(
-        ui.expanded.get("lead").is_some_and(|set| set.contains(&1)),
-        "ctrl+o expands the highlighted message"
-    );
-    let rev = ui.expanded_rev;
-    handle_deck_key(ctrl('o'), &model, &mut ui);
-    assert!(
-        ui.expanded.get("lead").is_none_or(|set| !set.contains(&1)),
-        "a second ctrl+o collapses it again"
-    );
-    assert!(
-        ui.expanded_rev > rev,
-        "each toggle invalidates the fold cache"
-    );
-}
-
-#[test]
-fn ctrl_o_with_no_selection_toggles_the_expand_all_overlay() {
-    let mut model = model_with(&["lead"]);
-    with_tool_exchange(&mut model, "lead");
-    let mut ui = ready_ui();
-
-    // First press (no selection): every expandable message opens at once —
-    // no per-entry set is touched.
-    handle_deck_key(ctrl('o'), &model, &mut ui);
-    assert!(ui.transcript_expand_all, "expand-all overlay on");
-    assert!(
-        ui.expanded.get("lead").is_none_or(|set| set.is_empty()),
-        "the overlay does not write the per-entry sets"
-    );
-    // Second press: everything closes again — ctrl+o is its own way out.
-    handle_deck_key(ctrl('o'), &model, &mut ui);
-    assert!(!ui.transcript_expand_all, "ctrl+o again collapses");
-}
-
-#[test]
-fn esc_collapses_the_expand_all_overlay_before_stopping_the_turn() {
-    let mut model = model_with(&["lead"]);
-    with_tool_exchange(&mut model, "lead"); // events flip the agent to Running
-    let mut ui = ready_ui();
-    handle_deck_key(ctrl('o'), &model, &mut ui);
-    assert!(ui.transcript_expand_all);
-
-    // Esc's first job here is the overlay — NOT cancelling the running
-    // turn (precedence rule 8 beats rules 9–10).
-    let action = handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-    assert!(!ui.transcript_expand_all, "esc is a graceful way out");
-    assert!(
-        matches!(action, DeckAction::Handled),
-        "the overlay-collapsing esc must not reach the turn-stop rules"
-    );
-    // With the overlay gone, the next Esc resumes normal duty (stop the
-    // in-flight turn).
-    let action = handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-    assert!(
-        matches!(action, DeckAction::Send(WorkspaceInput::Control { .. })),
-        "esc after the overlay closes stops the turn as before"
-    );
-}
-
-#[test]
-fn ctrl_o_on_a_highlight_peels_one_row_out_of_the_expand_all_overlay() {
-    let mut model = model_with(&["lead"]);
-    with_tool_exchange(&mut model, "lead"); // entries 0 (call) + 1 (result)
-    let mut ui = ready_ui();
-    handle_deck_key(ctrl('o'), &model, &mut ui); // overlay on
-    handle_deck_key(key(KeyCode::Up), &model, &mut ui); // highlight entry 1
-
-    handle_deck_key(ctrl('o'), &model, &mut ui);
-    assert!(
-        !ui.transcript_expand_all,
-        "the overlay materializes into per-entry expansions"
-    );
-    let set = ui.expanded.get("lead").expect("materialized set");
-    assert!(
-        set.contains(&0) && !set.contains(&1),
-        "the highlighted row collapsed; the rest stay open: {set:?}"
-    );
-}
-
-#[test]
-fn bracket_jumps_reach_both_ends_of_the_transcript() {
-    let mut model = model_with(&["lead"]);
-    with_tool_exchange(&mut model, "lead");
-    let mut ui = ready_ui();
-    // A scrollable transcript (metrics as the render pass would set them).
-    ui.metrics.session_total = 100;
-    ui.metrics.session_height = 10;
-
-    // ⌘/⌃ [ pins the window to the very beginning of the session…
-    handle_deck_key(ctrl('['), &model, &mut ui);
-    assert!(!ui.session_scroll.follow);
-    assert_eq!(ui.session_scroll.window(100, 10), 0..10, "jumped to start");
-
-    // …and ⌘/⌃ ] returns to the end, re-arming tail-follow. Both also
-    // drop any highlight so the pinned selection can't yank the view back.
-    let cmd_close = KeyEvent::new(KeyCode::Char(']'), KeyModifiers::SUPER);
-    handle_deck_key(cmd_close, &model, &mut ui);
-    assert!(ui.session_scroll.follow, "jumped to end = follow the tail");
-    assert_eq!(ui.session_selected, None);
-}
-
-#[test]
-fn down_past_the_last_message_clears_selection_and_rearms_follow() {
-    let mut model = model_with(&["lead"]);
-    with_tool_exchange(&mut model, "lead");
-    let mut ui = ready_ui();
-    handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-    handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-    assert_eq!(ui.session_selected, None, "down past the tail deselects");
-    assert!(ui.session_scroll.follow, "…and re-arms tail-follow");
-}
-
-#[test]
-fn ctrl_o_on_a_non_expandable_selection_is_a_no_op() {
-    let mut model = model_with(&["lead"]);
-    // A plain text message — `entry_lines` ignores the expanded flag for
-    // it, so ctrl+o has nothing to toggle.
-    model.apply_inbound(&Inbound::Event {
-        agent: "lead".into(),
-        event: AgentEvent::Text {
-            text: "hello".into(),
-        },
-    });
-    let mut ui = ready_ui();
-    handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-    assert_eq!(ui.session_selected, Some(0));
-
-    let rev = ui.expanded_rev;
-    handle_deck_key(ctrl('o'), &model, &mut ui);
-    assert!(
-        ui.expanded.get("lead").is_none_or(|set| set.is_empty()),
-        "nothing marked expanded"
-    );
-    assert_eq!(
-        ui.expanded_rev, rev,
-        "a no-op press must not invalidate the settled fold cache"
-    );
-}
-
-#[test]
-fn switching_focus_drops_the_session_selection() {
-    let mut model = model_with(&["lead", "sub"]);
-    with_tool_exchange(&mut model, "lead");
-    let mut ui = ready_ui();
-    handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-    assert!(ui.session_selected.is_some());
-
-    // Focus another agent from the Agents tab: the selection indexes the
-    // *previous* agent's transcript and must not carry across.
-    ui.tab = DeckTab::Agents;
-    handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-    assert_eq!(ui.focused, 1);
-    assert_eq!(
-        ui.session_selected, None,
-        "selection cleared on focus change"
-    );
-    assert!(ui.session_scroll.follow, "…and tail-follow re-arms");
-}
-
-#[test]
-fn deregister_keeps_focus_in_bounds_and_on_the_same_agent() {
-    let mut model = model_with(&["lead", "req:1", "req:2"]);
-    let mut ui = ready_ui();
-    ui.focused = 2; // req:2
-
-    // An EARLIER row vanishing shifts indexes down: the focused AGENT
-    // stays focused at its new index.
-    ingest_inbound(
-        &Inbound::Deregister {
-            agent: "req:1".into(),
-        },
-        &mut model,
-        &mut ui,
-    );
-    assert_eq!(ui.focused, 1);
-    assert_eq!(model.agents[ui.focused].meta.id, "req:2");
-
-    // The focused LAST row vanishing clamps focus back into range.
-    ingest_inbound(
-        &Inbound::Deregister {
-            agent: "req:2".into(),
-        },
-        &mut model,
-        &mut ui,
-    );
-    assert_eq!(ui.focused, 0, "focus stays in bounds");
-    assert_eq!(model.agents[ui.focused].meta.id, "lead");
-}
-
-#[test]
-fn deregister_of_the_focused_row_drops_the_stale_selection() {
-    let mut model = model_with(&["lead", "req:1", "req:2"]);
-    with_tool_exchange(&mut model, "req:1");
-    with_tool_exchange(&mut model, "req:2");
-    let mut ui = ready_ui();
-    ui.focused = 1; // req:1
-    ui.session_selected = Some(1); // a row of req:1's transcript
-
-    ingest_inbound(
-        &Inbound::Deregister {
-            agent: "req:1".into(),
-        },
-        &mut model,
-        &mut ui,
-    );
-    // The successor (req:2) slides into the focused index…
-    assert_eq!(ui.focused, 1);
-    assert_eq!(model.agents[1].meta.id, "req:2");
-    // …but the selection indexed the REMOVED agent's transcript, so it
-    // must not re-attach to the successor's (which is long enough that
-    // range-clamping alone would have kept it).
-    assert_eq!(ui.session_selected, None);
-    assert!(ui.session_scroll.follow, "tail-follow re-arms");
-}
-
 fn ready_ui() -> DeckUi {
     let mut ui = DeckUi::default();
     ui.splash.skip(); // past the splash for interaction tests
     ui
-}
-
-fn session_info(id: &str) -> crate::envelope::SessionInfo {
-    crate::envelope::SessionInfo {
-        id: id.into(),
-        title: format!("title for {id}"),
-        summary: String::new(),
-        workspace: "/tmp/w".into(),
-        phase: crate::envelope::SessionPhase::Complete,
-        started_ms: 0,
-        updated_ms: 0,
-        mine: false,
-        resumable: false,
-    }
-}
-
-fn notification(id: &str, read: bool, session: Option<&str>) -> crate::envelope::NotificationInfo {
-    crate::envelope::NotificationInfo {
-        id: id.into(),
-        title: "a title".into(),
-        body: "a body".into(),
-        source: String::new(),
-        created_ms: 0,
-        read,
-        session_id: session.map(str::to_string),
-    }
-}
-
-#[test]
-fn sessions_overlay_enter_opens_the_selected_session_and_closes() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ui.sessions_open = true;
-    ui.sessions = vec![session_info("ses-1"), session_info("ses-2")];
-    ui.sessions_sel = 1;
-
-    let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert_eq!(
-        action,
-        DeckAction::Send(WorkspaceInput::SessionOpen { id: "ses-2".into() }),
-        "⏎ opens (replays) the selected session"
-    );
-    assert!(!ui.sessions_open, "the overlay closes on open");
-}
-
-#[test]
-fn sessions_overlay_enter_with_no_rows_is_a_no_op() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ui.sessions_open = true; // registry snapshot empty
-
-    let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert_eq!(action, DeckAction::Handled, "nothing to open");
-    assert!(ui.sessions_open, "the overlay stays up (Esc closes it)");
-}
-
-#[test]
-fn inbox_enter_on_a_linked_notification_marks_read_and_opens_the_session() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ui.inbox_open = true;
-    ui.notifications = vec![notification("n1", false, Some("ses-9"))];
-
-    let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert_eq!(
-        action,
-        DeckAction::Send(WorkspaceInput::NotificationRead { id: "n1".into() }),
-        "the read goes out as the key's action"
-    );
-    assert_eq!(
-        ui.pending_inputs,
-        vec![WorkspaceInput::SessionOpen { id: "ses-9".into() }],
-        "…and the open rides pending_inputs right behind it"
-    );
-    assert!(!ui.inbox_open, "the overlay closes when a session opens");
-}
-
-#[test]
-fn inbox_enter_on_an_already_read_linked_notification_just_opens() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ui.inbox_open = true;
-    ui.notifications = vec![notification("n1", true, Some("ses-9"))];
-
-    let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert_eq!(
-        action,
-        DeckAction::Send(WorkspaceInput::SessionOpen { id: "ses-9".into() }),
-        "already read — no second NotificationRead, just the open"
-    );
-    assert!(ui.pending_inputs.is_empty());
-    assert!(!ui.inbox_open);
-}
-
-#[test]
-fn inbox_enter_without_a_session_link_keeps_the_mark_read_behavior() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ui.inbox_open = true;
-    ui.notifications = vec![notification("n1", false, None)];
-
-    let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert_eq!(
-        action,
-        DeckAction::Send(WorkspaceInput::NotificationRead { id: "n1".into() }),
-        "unlinked ⏎ is exactly the old mark-read"
-    );
-    assert!(ui.pending_inputs.is_empty(), "no session to open");
-    assert!(ui.inbox_open, "the overlay stays open, as before");
-
-    // Once read, ⏎ on an unlinked notification is a no-op.
-    ui.notifications = vec![notification("n1", true, None)];
-    let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert_eq!(action, DeckAction::Handled);
-    assert!(ui.inbox_open);
-}
-
-#[test]
-fn inbox_space_only_marks_read_and_never_opens() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ui.inbox_open = true;
-    ui.notifications = vec![notification("n1", false, Some("ses-9"))];
-
-    let action = handle_deck_key(key(KeyCode::Char(' ')), &model, &mut ui);
-    assert_eq!(
-        action,
-        DeckAction::Send(WorkspaceInput::NotificationRead { id: "n1".into() }),
-        "␣ keeps its plain mark-read meaning"
-    );
-    assert!(ui.pending_inputs.is_empty(), "␣ never opens the session");
-    assert!(ui.inbox_open, "␣ never closes the overlay");
-}
-
-#[test]
-fn eviction_clamps_the_selection_and_drops_stale_expansions() {
-    use crate::model::MAX_TRANSCRIPT_ENTRIES;
-    let mut model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    let retry = |i: usize| Inbound::Event {
-        agent: "lead".into(),
-        event: AgentEvent::Retry {
-            attempt: i as u32,
-            reason: "r".into(),
-        },
-    };
-    // Grow to just under the cap, then highlight + expand near the tail.
-    for i in 0..(MAX_TRANSCRIPT_ENTRIES - 1) {
-        ingest_inbound(&retry(i), &mut model, &mut ui);
-    }
-    ui.session_selected = Some(MAX_TRANSCRIPT_ENTRIES - 2);
-    toggle_expanded(&mut ui, "lead", MAX_TRANSCRIPT_ENTRIES - 2);
-    let rev = ui.expanded_rev;
-
-    // One more event crosses the cap: a chunk of the front evicts.
-    ingest_inbound(&retry(MAX_TRANSCRIPT_ENTRIES), &mut model, &mut ui);
-    let len = model.agents[0].model.transcript.len();
-    assert!(len < MAX_TRANSCRIPT_ENTRIES, "a chunk was evicted");
-    assert!(
-        ui.session_selected.is_some_and(|sel| sel < len),
-        "selection clamped into the retained window"
-    );
-    assert!(
-        !ui.expanded.contains_key("lead"),
-        "index-keyed expansions are stale once the front moved"
-    );
-    assert!(ui.expanded_rev > rev, "fold cache invalidated");
-}
-
-/// Front-eviction moves every retained index down, and the scrollback counter
-/// is an index. Off by even one and accessible mode either re-announces an
-/// entry a reader already heard or — worse — suppresses one from the live pane
-/// that was never written anywhere (#1258).
-///
-/// The assertion is exact rather than approximate: the entry the counter now
-/// points at must be the very next one after the last that was flushed.
-#[test]
-fn eviction_keeps_the_scrollback_counter_pointing_at_the_same_entry() {
-    use crate::model::{MAX_TRANSCRIPT_ENTRIES, TranscriptEntry};
-    let mut model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    let retry = |i: usize| Inbound::Event {
-        agent: "lead".into(),
-        event: AgentEvent::Retry {
-            attempt: i as u32,
-            reason: "r".into(),
-        },
-    };
-    for i in 0..(MAX_TRANSCRIPT_ENTRIES - 1) {
-        ingest_inbound(&retry(i), &mut model, &mut ui);
-    }
-    // Everything up to (not including) attempt 1000 is in scrollback — past
-    // the eviction chunk, so the survivors include flushed entries and the
-    // counter has real work to do (a prefix entirely inside the evicted chunk
-    // would correctly floor at zero and prove nothing about the arithmetic).
-    ui.scrollback.set_live(true);
-    ui.scrollback.record(&crate::accessible::FlushBlock {
-        agent: "lead".into(),
-        from: 0,
-        to: 1_000,
-        header: None,
-    });
-
-    ingest_inbound(&retry(MAX_TRANSCRIPT_ENTRIES), &mut model, &mut ui);
-    let transcript = &model.agents[0].model.transcript;
-    assert!(
-        transcript.len() < MAX_TRANSCRIPT_ENTRIES,
-        "a chunk was evicted"
-    );
-    let flushed = ui.scrollback.flushed_for("lead");
-    assert!(
-        matches!(&transcript[flushed], TranscriptEntry::Retry { attempt, .. } if *attempt == 1_000),
-        "the counter must still point one past the last flushed entry, got {:?}",
-        transcript[flushed]
-    );
-}
-
-#[test]
-fn any_key_dismisses_the_splash_first() {
-    let model = model_with(&["lead"]);
-    let mut ui = DeckUi::default(); // splash NOT skipped
-    assert!(!ui.splash.is_done());
-    assert_eq!(
-        handle_deck_key(ch('a'), &model, &mut ui),
-        DeckAction::Handled
-    );
-    assert!(ui.splash.is_done(), "first key skips the splash");
-    assert!(ui.composer.buffer().is_empty(), "and does not type");
-}
-
-#[test]
-fn splash_cues_replay_and_release_the_launch_mark() {
-    let mut model = WorkspaceModel::new();
-    let mut ui = DeckUi::default();
-    ui.splash.skip(); // the deck is up; `/init` arrives later
-    assert!(ui.splash.is_done());
-
-    // Replay: a fresh held mark owns the frame again for as long as init runs…
-    ingest_inbound(&Inbound::Splash(SplashCue::Replay), &mut model, &mut ui);
-    assert!(!ui.splash.is_done(), "replay re-holds the mark over init");
-
-    // …and Release bounds the mark's life: it plays the assemble out (the
-    // brand's one beat of screen time — exact timing is `splash::tests`)
-    // rather than cutting mid-build, while a key press still cuts at once.
-    ingest_inbound(&Inbound::Splash(SplashCue::Release), &mut model, &mut ui);
-    assert!(
-        !ui.splash.is_done(),
-        "a release straight after replay finishes the assemble first"
-    );
-    ui.splash.skip();
-    assert!(ui.splash.is_done(), "any key still cuts immediately");
-}
-
-/// The invariant this whole path exists for: the transcript is the home for
-/// agent and user messages ONLY. Session chrome used to ride an
-/// `AgentEvent::Text`, which auto-registered a lead lane and gave the notice a
-/// transcript row indistinguishable from the model speaking.
-#[test]
-fn a_system_notice_goes_to_the_dialog_and_never_to_the_transcript() {
-    let mut model = WorkspaceModel::new();
-    let mut ui = DeckUi::default();
-    ui.splash.skip();
-
-    let text = "◂ a previous session is resumable — ← opens SESSIONS";
-    ingest_inbound(&Inbound::Notice(text.to_string()), &mut model, &mut ui);
-
-    assert_eq!(ui.notice.entries(), [text], "the dialog holds the notice");
-    assert!(ui.notice.is_visible(), "and shows it");
-    assert!(
-        model.agents.is_empty(),
-        "a notice must not conjure an agent lane — an Event would have, and \
-         that lane's transcript is exactly where chrome does not belong"
-    );
-}
-
-/// Dismissal is total but not greedy: the notice goes, the keystroke lands.
-/// Swallowing it would eat the first character typed, in precisely the second
-/// or two when the dialog is up and the user is most likely to start typing.
-#[test]
-fn any_key_dismisses_the_startup_notice_without_eating_the_keystroke() {
-    let mut model = WorkspaceModel::new();
-    let mut ui = DeckUi::default();
-    ui.splash.skip();
-    ingest_inbound(
-        &Inbound::Notice("indexing…".to_string()),
-        &mut model,
-        &mut ui,
-    );
-    assert!(ui.notice.is_visible());
-
-    handle_deck_key(key(KeyCode::Char('a')), &model, &mut ui);
-    assert!(!ui.notice.is_visible(), "any key dismisses the dialog");
-    assert_eq!(
-        ui.composer.buffer(),
-        "a",
-        "and the keystroke still reaches the composer"
-    );
-}
-
-#[test]
-fn no_anim_sessions_ignore_splash_replays() {
-    let mut model = WorkspaceModel::new();
-    let mut ui = DeckUi::default();
-    ui.no_anim = true;
-    ui.splash.skip();
-    ingest_inbound(&Inbound::Splash(SplashCue::Replay), &mut model, &mut ui);
-    assert!(
-        ui.splash.is_done(),
-        "a no-anim session never re-holds the launch mark"
-    );
-}
-
-fn session_row(
-    id: &str,
-    phase: crate::envelope::SessionPhase,
-    mine: bool,
-    resumable: bool,
-) -> crate::envelope::SessionInfo {
-    crate::envelope::SessionInfo {
-        id: id.into(),
-        title: format!("ws: {id}"),
-        summary: String::new(),
-        workspace: "/w".into(),
-        phase,
-        started_ms: 0,
-        updated_ms: 0,
-        mine,
-        resumable,
-    }
-}
-
-#[test]
-fn sessions_overlay_enter_resumes_resumable_rows_and_opens_the_rest() {
-    use crate::envelope::SessionPhase;
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ui.sessions_open = true;
-    ui.sessions = vec![
-        session_row("ses-mine", SessionPhase::InProgress, true, false),
-        session_row("ses-paused", SessionPhase::Paused, false, true),
-        session_row("ses-foreign", SessionPhase::Complete, false, false),
-    ];
-
-    // Grouped order: InProgress (mine) · Paused (resumable) · Complete.
-    // ⏎ on the resumable row navigates into it LIVE: the overlay closes
-    // and the driver is told to resume exactly that session.
-    ui.sessions_sel = 1;
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::SessionResume {
-            id: "ses-paused".into()
-        })
-    );
-    assert!(!ui.sessions_open, "the overlay closes on navigation");
-
-    // ⏎ on any non-resumable row — this deck's own included — opens a
-    // read-only replay instead (the `replay:<id>` lane).
-    for (sel, id) in [(0, "ses-mine"), (2, "ses-foreign")] {
-        ui.sessions_open = true;
-        ui.sessions_sel = sel;
-        assert_eq!(
-            handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-            DeckAction::Send(WorkspaceInput::SessionOpen { id: id.into() })
-        );
-        assert!(!ui.sessions_open, "the overlay closes on open too");
-    }
-}
-
-#[test]
-fn paused_sessions_group_between_needs_input_and_cancelled() {
-    use crate::envelope::SessionPhase;
-    let mut ui = DeckUi::default();
-    ui.sessions = vec![
-        session_row("c", SessionPhase::Cancelled, false, true),
-        session_row("p", SessionPhase::Paused, false, true),
-        session_row("n", SessionPhase::NeedsInput, false, false),
-    ];
-    let order: Vec<&str> = grouped_session_rows(&ui)
-        .iter()
-        .map(|s| s.id.as_str())
-        .collect();
-    assert_eq!(order, ["n", "p", "c"]);
-}
-
-#[test]
-fn only_tab_switches_tabs_and_digits_always_type() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    assert_eq!(ui.tab, DeckTab::Session);
-    handle_deck_key(key(KeyCode::Tab), &model, &mut ui);
-    assert_eq!(ui.tab, DeckTab::Agents);
-    handle_deck_key(key(KeyCode::BackTab), &model, &mut ui);
-    assert_eq!(ui.tab, DeckTab::Session);
-    // A digit with an empty composer starts the prompt — it never jumps
-    // to a tab, so prompts can begin with 1–5.
-    handle_deck_key(ch('3'), &model, &mut ui);
-    assert_eq!(ui.tab, DeckTab::Session, "digit typed, tab unchanged");
-    handle_deck_key(ch('h'), &model, &mut ui);
-    handle_deck_key(ch('2'), &model, &mut ui);
-    assert_eq!(ui.composer.buffer(), "3h2");
-}
-
-#[test]
-fn bare_enter_always_enqueues_a_prompt_without_blocking() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    for c in "do the thing".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert_eq!(
-        action,
-        DeckAction::Send(WorkspaceInput::Enqueue {
-            text: "do the thing".into()
-        })
-    );
-    assert!(
-        ui.composer.buffer().is_empty(),
-        "composer clears after submit"
-    );
-}
-
-#[test]
-fn a_modified_enter_inserts_a_line_break_preserved_through_submit() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    for c in "line one".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    assert_eq!(
-        handle_deck_key(cmd_enter(), &model, &mut ui),
-        DeckAction::Handled,
-        "⌘⏎ is a line break, not a submit"
-    );
-    for c in "line two".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert_eq!(
-        action,
-        DeckAction::Send(WorkspaceInput::Enqueue {
-            text: "line one\nline two".into()
-        }),
-        "the typed line break survives into the submitted prompt"
-    );
-}
-
-#[test]
-fn plain_enter_on_a_blank_composer_inserts_nothing() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ui.tab = DeckTab::Graph; // a tab with no Enter binding of its own
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-        DeckAction::Ignored
-    );
-    assert!(ui.composer.buffer().is_empty(), "no stray leading newline");
-}
-
-#[test]
-fn alt_brackets_jump_the_cursor_to_start_and_end() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    for c in "abc".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    assert_eq!(ui.composer.cursor(), 3);
-    assert_eq!(
-        handle_deck_key(alt('['), &model, &mut ui),
-        DeckAction::Handled
-    );
-    assert_eq!(ui.composer.cursor(), 0, "⌥[ → before the first character");
-    assert_eq!(
-        handle_deck_key(alt(']'), &model, &mut ui),
-        DeckAction::Handled
-    );
-    assert_eq!(ui.composer.cursor(), 3, "⌥] → one past the last character");
-}
-
-#[test]
-fn bare_enter_queues_and_a_modified_enter_inserts_a_break() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    for c in "hi".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    let alt_enter = KeyEvent::new(KeyCode::Enter, KeyModifiers::ALT);
-    assert_eq!(
-        handle_deck_key(alt_enter, &model, &mut ui),
-        DeckAction::Handled,
-        "⌥⏎ inserts a line break"
-    );
-    assert_eq!(ui.composer.buffer(), "hi\n");
-    let action = handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert_eq!(
-        action,
-        DeckAction::Send(WorkspaceInput::Enqueue {
-            text: "hi\n".into()
-        }),
-        "bare ⏎ queues (never blocks)"
-    );
-}
-
-#[test]
-fn arrow_keys_edit_a_multiline_prompt_instead_of_scrolling() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    for c in "ab".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    handle_deck_key(cmd_enter(), &model, &mut ui); // ⌘⏎ inserts a line break
-    for c in "cd".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    // ↑ moves the cursor into the first line (not the session scroll,
-    // and NOT the queue editor — the composer is not empty).
-    handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-    assert_eq!(ui.composer.cursor(), 2, "column kept on the line above");
-    handle_deck_key(key(KeyCode::Left), &model, &mut ui);
-    handle_deck_key(ch('X'), &model, &mut ui);
-    assert_eq!(ui.composer.buffer(), "aXb\ncd", "typed at the cursor");
-}
-
-#[test]
-fn ctrl_c_quits_from_any_tab() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ui.tab = DeckTab::Graph;
-    assert_eq!(
-        handle_deck_key(ctrl('c'), &model, &mut ui),
-        DeckAction::Quit
-    );
-}
-
-/// A one-agent model whose lead is mid-turn (`Running`).
-fn running_model() -> WorkspaceModel {
-    let mut m = model_with(&["lead"]);
-    m.apply_inbound(&Inbound::Event {
-        agent: "lead".into(),
-        event: AgentEvent::Text {
-            text: "working".into(),
-        },
-    });
-    m
-}
-
-/// The single-Esc outcome: a clean stop for the lead's in-flight turn.
-fn stop_lead() -> DeckAction {
-    DeckAction::Send(WorkspaceInput::Control {
-        agent: "lead".into(),
-        control: AgentControl::Stop,
-    })
-}
-
-#[test]
-fn esc_stops_a_running_turn_and_arms_the_double_press() {
-    let model = running_model();
-    let mut ui = ready_ui();
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        stop_lead()
-    );
-    assert!(
-        ui.esc_armed_at.is_some(),
-        "the stop arms the double-Esc window"
-    );
-}
-
-#[test]
-fn esc_with_no_turn_running_stays_inert() {
-    let model = model_with(&["lead"]); // Queued — nothing in flight
-    let mut ui = ready_ui();
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Ignored,
-        "an idle Esc must not send a stray stop"
-    );
-    assert!(ui.esc_armed_at.is_none());
-}
-
-/// **The steering witness.** Esc with something typed *delivers* it into the
-/// running turn and lets the turn continue.
-///
-/// Esc used to cancel unconditionally, which on the staged pipeline is not an
-/// interrupt but a restart: the driver truncates the partial turn and
-/// re-dispatches the backlog as a new one, re-entering at triage and context
-/// recall. So a user who typed one correcting sentence and pressed Esc paid
-/// for the whole turn again to deliver it — and the sentence itself stayed
-/// stranded in the composer while the turn it was written for died.
-#[test]
-fn esc_with_a_typed_draft_steers_the_running_turn_instead_of_cancelling_it() {
-    let model = running_model();
-    let mut ui = ready_ui();
-    for c in "use the existing helper".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::Steer {
-            agent: "lead".into(),
-            texts: vec!["use the existing helper".into()],
-        }),
-        "the draft is delivered to the turn in flight, not thrown away with it"
-    );
-    assert_eq!(
-        ui.composer.buffer(),
-        "",
-        "the draft was DELIVERED, so it must not also survive as a next-turn \
-         prompt — the same words dispatched twice is its own defect"
-    );
-    // The double-Esc escalation is untouched: a full stop is still one more
-    // press away, which is what keeps "I want it to stop" reachable.
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::StopAndHold {
-            agent: "lead".into()
-        })
-    );
-}
-
-/// The queue half: every waiting prompt goes in, in order, in one message —
-/// which is the behaviour the request described ("whatever is in my prompt
-/// queue, even if there is more than one, gets sent in and the agent keeps
-/// working").
-///
-/// One message rather than N, deliberately: draining the queue as separate
-/// inputs would let a second Esc land between two of them and cancel the turn
-/// holding the rest.
-#[test]
-fn esc_delivers_the_whole_prompt_queue_in_order() {
-    let mut model = running_model();
-    model.queue.enqueue("first correction".into(), 0);
-    model.queue.enqueue("> second correction".into(), 1);
-    let mut ui = ready_ui();
-    for c in "and the draft".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::Steer {
-            agent: "lead".into(),
-            texts: vec![
-                "first correction".into(),
-                // The `>` marker is stripped: it already means "steer", and
-                // this whole path is steering.
-                "second correction".into(),
-                // The draft is newest, so it is read last.
-                "and the draft".into(),
-            ],
-        })
-    );
-}
-
-/// And the case where Esc still means stop: nothing typed, nothing queued.
-/// With no guidance to deliver, an interrupt is the only thing the keystroke
-/// can mean — so the escape hatch never moved.
-#[test]
-fn esc_with_nothing_to_say_is_still_a_plain_interrupt() {
-    let model = running_model();
-    let mut ui = ready_ui();
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        stop_lead()
-    );
-}
-
-#[test]
-fn double_esc_inside_the_window_escalates_to_stop_and_hold() {
-    let model = running_model();
-    let mut ui = ready_ui();
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        stop_lead()
-    );
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::StopAndHold {
-            agent: "lead".into()
-        })
-    );
-    assert!(
-        ui.dispatch_held,
-        "the deck now front-inserts its next submission"
-    );
-    assert!(
-        ui.esc_armed_at.is_none(),
-        "the pair resets after escalating"
-    );
-}
-
-#[test]
-fn the_second_esc_fires_even_if_the_cancel_already_folded() {
-    // Between the two presses the first cancel's error event may fold
-    // (status `Failed`) before the auto-dispatched next prompt produces
-    // any event — the escalation must not be lost to that gap.
-    let mut model = running_model();
-    let mut ui = ready_ui();
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        stop_lead()
-    );
-    model.apply_inbound(&Inbound::Event {
-        agent: "lead".into(),
-        event: AgentEvent::Error {
-            message: "turn stopped by user".into(),
-            retryable: false,
-        },
-    });
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::StopAndHold {
-            agent: "lead".into()
-        })
-    );
-}
-
-#[test]
-fn an_intervening_key_breaks_the_double_esc_pair() {
-    let model = running_model();
-    let mut ui = ready_ui();
-    handle_deck_key(key(KeyCode::Esc), &model, &mut ui);
-    handle_deck_key(ch('x'), &model, &mut ui); // types into the composer
-    assert!(ui.esc_armed_at.is_none(), "any other key disarms");
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::Steer {
-            agent: "lead".into(),
-            texts: vec!["x".into()],
-        }),
-        "the next Esc is a fresh single press, not an escalation — and with a \
-         draft in the composer a single press steers"
-    );
-}
-
-#[test]
-fn a_stale_arm_outside_the_window_does_not_escalate() {
-    let model = running_model();
-    let mut ui = ready_ui();
-    // Backdate the arm past the window. (If the monotonic clock is too
-    // young to backdate, `checked_sub` leaves it unarmed — which expects
-    // the same single-stop outcome.)
-    ui.esc_armed_at = std::time::Instant::now()
-        .checked_sub(ESC_DOUBLE_WINDOW + std::time::Duration::from_secs(1));
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        stop_lead(),
-        "past the window, Esc is a single stop again"
-    );
-}
-
-#[test]
-fn esc_dismisses_the_slash_popup_instead_of_stopping_the_turn() {
-    let model = running_model();
-    let mut ui = ready_ui();
-    ui.slash_commands = vec![SlashCommand::new("/help", "help")];
-    handle_deck_key(ch('/'), &model, &mut ui);
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Handled,
-        "rule 4: the popup claims Esc — no stop is sent"
-    );
-    assert!(ui.esc_armed_at.is_none(), "a claimed Esc never arms");
-}
-
-#[test]
-fn an_esc_claimed_by_the_queue_editor_breaks_the_pair_too() {
-    let mut model = model_with_queue(&["one"]);
-    model.apply_inbound(&Inbound::Event {
-        agent: "lead".into(),
-        event: AgentEvent::Text { text: "hi".into() },
-    });
-    let mut ui = ready_ui();
-    ui.queue_open = true;
-    ui.esc_armed_at = Some(std::time::Instant::now());
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Handled,
-        "rule 3: the queue editor claims Esc — no stop is sent"
-    );
-    assert!(!ui.queue_open, "…it closed the editor");
-    assert!(ui.esc_armed_at.is_none(), "…and broke the double-Esc pair");
-}
-
-#[test]
-fn esc_closes_an_open_diff_before_it_stops_the_turn() {
-    let model = running_model();
-    let mut ui = ready_ui();
-    ui.tab = DeckTab::Files;
-    ui.files_diff_open = true;
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Handled,
-        "rule 6: the diff overlay claims Esc"
-    );
-    assert!(!ui.files_diff_open);
-}
-
-#[test]
-fn esc_clears_a_session_highlight_before_it_stops_the_turn() {
-    let mut model = running_model();
-    with_tool_exchange(&mut model, "lead");
-    let mut ui = ready_ui();
-    handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-    assert!(ui.session_selected.is_some());
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        DeckAction::Handled,
-        "rule 7: the highlight claims Esc first"
-    );
-    assert_eq!(ui.session_selected, None);
-    // With nothing left to claim it, the NEXT Esc stops the turn (and
-    // since the claimed Esc broke the pair, it is a single stop).
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Esc), &model, &mut ui),
-        stop_lead()
-    );
-}
-
-#[test]
-fn the_first_submission_after_a_hold_enqueues_at_the_front() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ui.dispatch_held = true;
-    for c in "urgent fix".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::EnqueueFront {
-            text: "urgent fix".into()
-        }),
-        "the held submission jumps the queue"
-    );
-    assert!(!ui.dispatch_held, "the submission releases the hold");
-    for c in "later".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::Enqueue {
-            text: "later".into()
-        }),
-        "after the hold clears, submissions append as usual"
-    );
-}
-
-#[test]
-fn agents_tab_controls_fire_only_when_composer_empty() {
-    let model = model_with(&["lead", "sub"]);
-    let mut ui = ready_ui();
-    ui.tab = DeckTab::Agents;
-    ui.focused = 1;
-    // 's' with empty composer → Stop control for the focused agent.
-    assert_eq!(
-        handle_deck_key(ch('s'), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::Control {
-            agent: "sub".into(),
-            control: AgentControl::Stop,
-        })
-    );
-    // With text typed, 's' types instead.
-    handle_deck_key(ch('h'), &model, &mut ui);
-    handle_deck_key(ch('s'), &model, &mut ui);
-    assert_eq!(ui.composer.buffer(), "hs");
-}
-
-#[test]
-fn agents_updown_moves_focus_and_enter_opens_session() {
-    let model = model_with(&["a", "b", "c"]);
-    let mut ui = ready_ui();
-    ui.tab = DeckTab::Agents;
-    handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-    handle_deck_key(key(KeyCode::Down), &model, &mut ui);
-    assert_eq!(ui.focused, 2);
-    handle_deck_key(key(KeyCode::Up), &model, &mut ui);
-    assert_eq!(ui.focused, 1);
-    handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert_eq!(ui.tab, DeckTab::Session);
-}
-
-#[test]
-fn ask_user_answer_latches_until_a_fresh_question_rearms() {
-    let ask = Inbound::Event {
-        agent: "lead".into(),
-        event: AgentEvent::AskUser {
-            id: "call_ask_1".into(),
-            question: "which db?".into(),
-            options: vec!["postgres".into(), "sqlite".into()],
-        },
-    };
-    let mut model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    ingest_inbound(&ask, &mut model, &mut ui);
-
-    assert_eq!(
-        handle_deck_key(ch('1'), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::ToAgent {
-            agent: "lead".into(),
-            input: UserInput::AskUserAnswer {
-                id: "call_ask_1".into(),
-                answer: "postgres".into(),
-            },
-        })
-    );
-    // The question stays pending until its ToolResult lands; a second digit
-    // must type into the composer, not answer the question again.
-    assert!(model.agents[0].model.pending_ask_user.is_some());
-    handle_deck_key(ch('2'), &model, &mut ui);
-    assert_eq!(
-        ui.composer.buffer(),
-        "2",
-        "second press types, never re-sends"
-    );
-
-    // A FRESH question re-arms the quick-pick.
-    handle_deck_key(key(KeyCode::Backspace), &model, &mut ui);
-    ingest_inbound(&ask, &mut model, &mut ui);
-    assert_eq!(
-        handle_deck_key(ch('2'), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::ToAgent {
-            agent: "lead".into(),
-            input: UserInput::AskUserAnswer {
-                id: "call_ask_1".into(),
-                answer: "sqlite".into(),
-            },
-        }),
-        "a new question re-arms the quick-pick"
-    );
-}
-
-/// The routing card owns every key ahead of the composer once raised — but
-/// it used to have no way out if the turn it was asking about finished (or
-/// errored, or itself started asking a question) before the user answered
-/// `s`/`n`/`p`/Esc: every further keystroke, including Enter and Backspace,
-/// died silently at the card's catch-all. This is the "can't clear or submit
-/// anything after a turn completes" bug — the fix releases the card the same
-/// way Esc does the moment its agent stops being `Running`.
-#[test]
-fn a_finished_turn_releases_a_routing_card_stuck_on_its_prompt() {
-    let mut model = model_with(&["lead"]);
-    model.apply_inbound(&Inbound::Status {
-        agent: "lead".into(),
-        status: crate::AgentStatus::Running,
-    });
-    let mut ui = ready_ui();
-
-    for c in "and now the tests".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-        DeckAction::Handled,
-        "an ambiguous mid-turn prompt raises the card instead of sending"
-    );
-    assert!(ui.pending_dispatch.is_some(), "the card is up");
-
-    ingest_inbound(
-        &Inbound::Event {
-            agent: "lead".into(),
-            event: AgentEvent::RunComplete {
-                model: "test-model".into(),
-                cost_usd: 0.0,
-            },
-        },
-        &mut model,
-        &mut ui,
-    );
-
-    assert!(
-        ui.pending_dispatch.is_none(),
-        "a finished turn releases the card rather than leaving it stuck"
-    );
-    assert_eq!(
-        ui.composer.buffer(),
-        "and now the tests",
-        "the held prompt returns to the composer, same as Esc"
-    );
-
-    // The keyboard is unstuck: the same text submits normally now, since
-    // `dispatch::route`'s own `running` check is false for an idle agent.
-    assert_eq!(
-        handle_deck_key(key(KeyCode::Enter), &model, &mut ui),
-        DeckAction::Send(WorkspaceInput::Enqueue {
-            text: "and now the tests".into()
-        })
-    );
-}
-
-/// The release is scoped to the agent the card was raised against — another
-/// agent finishing its own turn must not discard a card still legitimately
-/// waiting on the user.
-#[test]
-fn a_finished_turn_on_another_agent_does_not_release_someone_elses_card() {
-    let mut model = model_with(&["lead"]);
-    model.apply_inbound(&Inbound::Status {
-        agent: "lead".into(),
-        status: crate::AgentStatus::Running,
-    });
-    let mut ui = ready_ui();
-
-    for c in "keep me".chars() {
-        handle_deck_key(ch(c), &model, &mut ui);
-    }
-    handle_deck_key(key(KeyCode::Enter), &model, &mut ui);
-    assert!(ui.pending_dispatch.is_some());
-
-    ingest_inbound(
-        &Inbound::Event {
-            agent: "req:2".into(),
-            event: AgentEvent::TurnComplete {
-                model: "test-model".into(),
-                cost_usd: 0.0,
-            },
-        },
-        &mut model,
-        &mut ui,
-    );
-
-    assert!(
-        ui.pending_dispatch.is_some(),
-        "a different agent's completion must not release this card"
-    );
-}
-
-#[test]
-fn tab_and_backtab_walk_the_tab_bar() {
-    let model = model_with(&["lead"]);
-    let mut ui = ready_ui();
-    assert_eq!(ui.tab, DeckTab::Session);
-
-    handle_deck_key(key(KeyCode::Tab), &model, &mut ui);
-    assert_eq!(ui.tab, DeckTab::Agents);
-    handle_deck_key(key(KeyCode::BackTab), &model, &mut ui);
-    assert_eq!(ui.tab, DeckTab::Session);
-
-    // Re-selecting the active tab is a no-op, not an error.
-    ui.set_tab(DeckTab::Session);
-    assert_eq!(ui.tab, DeckTab::Session);
-}
-
-#[test]
-fn traces_filter_cycles_through_agents_and_back() {
-    let model = model_with(&["a", "b"]);
-    let mut ui = ready_ui();
-    ui.tab = DeckTab::Traces;
-    assert_eq!(ui.trace_filter, None);
-    handle_deck_key(ch('f'), &model, &mut ui);
-    assert_eq!(ui.trace_filter.as_deref(), Some("a"));
-    handle_deck_key(ch('f'), &model, &mut ui);
-    assert_eq!(ui.trace_filter.as_deref(), Some("b"));
-    handle_deck_key(ch('f'), &model, &mut ui);
-    assert_eq!(ui.trace_filter, None);
 }
